// src/app/(root)/ctgom/interface.ts
export interface Municipio {
  id: number;
  provincia_id: number;
  nome: string;
  deleted_at: string | null;
}

export interface Provincia {
  id: number;
  nome: string;
  criado_por: number;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
  municipios: Municipio[]; // tipagem do array de municípios
}

export interface Membro {
  id: number;
  nome: string;
  funcao: string;
  contacto: string;
  email: string;
  imagem: string;
}

export interface Mandato_ {
  id: number;
  data_inicio: string;
  data_fim: string | null;
  ctgom_id: number;
  membros: Membro[];
}



export interface DtoResponseGetAllCtgom {
  success: boolean;
  dados: Ctgom[];
  mensagem: string;
}

export interface DtoResponseGetCtgom {
  success: boolean;
  dados: Ctgom;
  mensagem: string;
}
export type DocumentoTipo = "imagem" | "video" | "pdf" | "outro";

export interface CtgomDocumento {
  id: number;
  ctgom_id: number;
  tipo: DocumentoTipo;
  nome: string;
  mandato_id: number;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
  url_ficheiro: string;
}
// ...
export interface Ctgom {
  id: number;
  nome: string;
  codigo: string;
  localizacao: string | null;
  ano: string; // formato YYYY-MM-DD
  municipio_id: number;
  criado_por: Usuario;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
  municipio: Municipio;
  documento: CtgomDocumento[];
  mandato: Mandato[];
 
}

export interface Usuario {
  id: number;
  nome: string;
  email: string;
  telefone: string;
  bi: string;
  data_nascimento: string; // "YYYY-MM-DD"
  genero: string;
  status: string;
  tipo: string;
  natural_id: number | null;
  criado_por: number | null;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
}

export interface Municipio {
  id: number;
  nome: string;
  provincia_id: number;
  criado_por: number;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
   provincia: Provincia;
}

export interface Documento {
  // estrutura ainda não definida no JSON
}
// ...
export interface Mandato {
  id: number
  ctgom_id: number
  data_inicio: string
  data_fim: string | null
  created_at: string
  updated_at: string
  deleted_at: string | null
  membros: Membro[]
  documentos: Documento[]
 
}